Added some useful functions.
-Added a trigger Function that emulates the triggerzone but you can just
call it any time.
Added a new <Trigger> node in the xml that can be used to Trigger
another object in the xml.
Changed how the interactive things call the text box now you just
trigger an object and if it finds a some text it will display it if all
its conditions are true.