textbox change
added slight pauses when using punctuation in the textbox. changed the
text from displaying every frame to a delta time event. (so now pauses
work by added to a timer when it reads a certain letter)
now we can use things like "..." at end of sentances and it will make
pauses for them. I think it looks better than before